Upstream: package refactor
